refactor: adding client only to rendering in client
